BE: [refactor] hospital response dto 디렉토리 수정 #73
